Moved module_utils.py into dependency.py
